src/workflow/evaluation_system.py
"""
Workflow structure and evaluation system for the prompt engineering framework.
"""

class WorkflowStage:
    def __init__(self, name, description, required_agents, evaluation_criteria):
        self.name = name
        self.description = description
        self.required_agents = required_agents
        self.evaluation_criteria = evaluation_criteria
        self.feedback_loops = []
        
    def add_feedback_loop(self, source_stage, criteria):
        """Add a feedback loop from another stage."""
        self.feedback_loops.append({
            "source": source_stage,
            "criteria": criteria
        })

class WorkflowManager:
    def __init__(self):
        self.stages = self._initialize_stages()
        self._setup_feedback_loops()
        
    def _initialize_stages(self):
        return {
            "problem_analysis": WorkflowStage(
                name="Problem Analysis",
                description="Initial problem understanding and context gathering",
                required_agents=["domain", "technical", "user"],
                evaluation_criteria={
                    "completeness": "Problem scope fully defined",
                    "clarity": "Clear objectives and constraints",
                    "context": "Relevant context captured",
                    "stakeholders": "All stakeholders identified"
                }
            ),
            "solution_ideation": WorkflowStage(
                name="Solution Ideation",
                description="Creative solution generation and brainstorming",
                required_agents=["creative", "technical", "domain"],
                evaluation_criteria={
                    "innovation": "Novel and creative approaches",
                    "feasibility": "Technical implementation possible",
                    "effectiveness": "Addresses core problems",
                    "scalability": "Solution can scale"
                }
            ),
            "technical_assessment": WorkflowStage(
                name="Technical Assessment",
                description="Detailed technical analysis and planning",
                required_agents=["technical", "data", "risk"],
                evaluation_criteria={
                    "architecture": "Sound technical design",
                    "integration": "Clear integration points",
                    "performance": "Performance requirements met",
                    "security": "Security considerations addressed"
                }
            ),
            "implementation_strategy": WorkflowStage(
                name="Implementation Strategy",
                description="Detailed execution planning",
                required_agents=["technical", "qa", "user"],
                evaluation_criteria={
                    "timeline": "Realistic implementation timeline",
                    "resources": "Required resources identified",
                    "risks": "Risk mitigation strategies",
                    "milestones": "Clear progress indicators"
                }
            )
        }
    
    def _setup_feedback_loops(self):
        """Initialize feedback loops between stages."""
        # Feedback loop 1: Implementation insights back to solution ideation
        self.stages["solution_ideation"].add_feedback_loop(
            "implementation_strategy",
            ["feasibility_updates", "resource_constraints"]
        )
        
        # Feedback loop 2: Technical findings back to problem analysis
        self.stages["problem_analysis"].add_feedback_loop(
            "technical_assessment",
            ["technical_constraints", "architecture_implications"]
        )
        
        # Feedback loop 3: User feedback to solution refinement
        self.stages["solution_ideation"].add_feedback_loop(
            "problem_analysis",
            ["user_needs", "requirement_updates"]
        )
        
        # Feedback loop 4: Quality metrics to implementation
        self.stages["implementation_strategy"].add_feedback_loop(
            "technical_assessment",
            ["quality_metrics", "performance_requirements"]
        )

class EvaluationSystem:
    def __init__(self):
        self.criteria = {
            "technical_viability": {
                "weight": 0.2,
                "metrics": [
                    "architecture_soundness",
                    "scalability_potential",
                    "integration_complexity"
                ]
            },
            "innovation_potential": {
                "weight": 0.15,
                "metrics": [
                    "novelty_score",
                    "market_differentiation",
                    "future_adaptability"
                ]
            },
            "user_impact": {
                "weight": 0.2,
                "metrics": [
                    "user_experience",
                    "accessibility",
                    "value_delivery"
                ]
            },
            "ethical_considerations": {
                "weight": 0.15,
                "metrics": [
                    "fairness_score",
                    "bias_assessment",
                    "transparency_level"
                ]
            },
            "implementation_feasibility": {
                "weight": 0.15,
                "metrics": [
                    "resource_requirements",
                    "timeline_realism",
                    "technical_complexity"
                ]
            },
            "risk_assessment": {
                "weight": 0.15,
                "metrics": [
                    "security_vulnerabilities",
                    "compliance_issues",
                    "operational_risks"
                ]
            }
        }
    
    def evaluate_solution(self, solution_data):
        """Evaluate a solution against all criteria."""
        scores = {}
        for criterion, details in self.criteria.items():
            scores[criterion] = self._evaluate_criterion(
                solution_data,
                details["metrics"],
                details["weight"]
            )
        return scores
    
    def _evaluate_criterion(self, solution_data, metrics, weight):
        """Evaluate a specific criterion based on its metrics."""
        # Implementation for metric evaluation
        pass
